+++ spec/rubocop/cop/rails/reflection_class_name_spec.rb
@@ -59,6 +59,12 @@ RSpec.describe RuboCop::Cop::Rails::ReflectionClassName, :config do
     RUBY
   end
 
+  it 'does not register an offense when using `class_name: to_s`' do
+    expect_no_offenses(<<~'RUBY')
+      has_many :accounts, class_name: to_s
+    RUBY
+  end
+
   it 'does not register an offense when using `foreign_key :account_id`' do
     expect_no_offenses(<<~RUBY)
       has_many :accounts, class_name: 'Account', foreign_key: :account_id